Move Frame and ReplClassLoader to util module
So that they can be used from the compiler module
