+++ amm/runtime/src/main/scala/ammonite/runtime/ClassLoaders.scala
@@ -25,7 +25,7 @@ class Frame(val classloader: SpecialClassLoader,
             val pluginClassloader: SpecialClassLoader,
             private[this] var imports0: Imports,
             private[this] var classpath0: Seq[java.net.URL],
-            private[this] var usedEarlierDefinitions0: Seq[String]) extends ammonite.repl.api.Frame{
+            private[this] var usedEarlierDefinitions0: Seq[String]) extends ammonite.util.Frame{
   private var frozen0 = false
   def frozen = frozen0
   def freeze(): Unit = {
@@ -204,7 +204,7 @@ class SpecialClassLoader(parent: ClassLoader,
                          parentSignature: Seq[(Either[String, java.net.URL], Long)],
                          var specialLocalClasses: Set[String],
                          urls: URL*)
-  extends ammonite.repl.api.ReplClassLoader(urls.toArray, parent){
+  extends ammonite.util.ReplClassLoader(urls.toArray, parent){
 
   /**
     * Files which have been compiled, stored so that our special
